+++ src/etheroll/utils.py
@@ -9,6 +9,7 @@ from kivy.app import App
 from kivy.clock import mainthread
 from kivy.lang import Builder
 from kivy.metrics import dp
+from kivy.storage.jsonstore import JsonStore
 from kivy.uix.screenmanager import Screen
 from kivy.utils import platform
 from kivymd.dialog import MDDialog
@@ -236,3 +237,24 @@ class SubScreen(Screen):
         """
         app = App.get_running_app()
         app.root.ids.toolbar_id.load_default_buttons()
+
+
+class Store:
+
+    @staticmethod
+    def get_store_path():
+        """
+        Returns the full user store path.
+        """
+        user_data_dir = App.get_running_app().user_data_dir
+        store_path = os.path.join(user_data_dir, 'store.json')
+        return store_path
+
+    @classmethod
+    def get_store(cls):
+        """
+        Returns the full user Store object instance.
+        """
+        store_path = cls.get_store_path()
+        store = JsonStore(store_path)
+        return store